Update #1
Hey the client began to work!
Right now, the clientside allows for player movement. the little rectangle moves along with the player's "a" and "d" keys. The platforms do need to be correctly sized, and a few FPS isses were fixed. The current state is extremely rough, but hopefully more of the movement can be completed by tomorrow.